+++ source/renderer/zettlr-body.js
@@ -109,14 +109,18 @@ class ZettlrBody {
     $(document).on('keydown', (event) => {
       let isDarwin = $('body').hasClass('darwin')
       let cmdOrCtrl = (isDarwin && event.metaKey) || (!isDarwin && event.ctrlKey)
-      let focusEditorShortcut = (cmdOrCtrl && event.shiftKey && event.which === 69)
-      let focusSidebarShortcut = (cmdOrCtrl && event.shiftKey && event.which === 84)
+      console.log(event.key)
+      let focusEditorShortcut = (cmdOrCtrl && event.shiftKey && event.key === 'e')
+      let focusSidebarShortcut = (cmdOrCtrl && event.shiftKey && event.key === 't')
       if (focusEditorShortcut) { // Cmd/Ctrl+Shift+E
         // Obviously, focus the editor
         this._renderer.getEditor().getEditor().focus()
       } else if (focusSidebarShortcut) { // Cmd/Ctrl+Shift+T
         // You know what to do
         $('#file-list').focus()
+      } else if (event.key === 'F2') {
+        // Trigger a rename
+        this.requestNewFileName(this._renderer.getActiveFile())
       }
     })